admin-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
asyncpg==0.28.0
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0.post2
boto3==1.28.38
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
